+++ docs/changes.md
@@ -37,6 +37,7 @@
 - add callbacks for cutters in #1020 and `examples/volumetric/slice_plane3.py`
 - add `utils.andrews_curves()` function.
 - add `transformations.LinearTransform.transpose()` method.
+- add `pointcloud.generate_segments()` to generate a continous line from un-ordered points in 3d
 
 
 ## Breaking changes
@@ -61,6 +62,7 @@
 ```
 examples/advanced/warp4b.py
 examples/advanced/diffuse_data.py
+examples/advanced/moving_least_squares1D.py
 
 examples/volumetric/slab_vol.py
 examples/volumetric/streamlines1.py